Major optmizations (READ DESC!)
Entity.cpp removed
units and buildings stats moved to Entitiy.h
new methods for IsUnit(), IsHero() and so
merged unitDB and buildingDB into a single DataBase(Entity*)
unitType, heroType and buildingType merged into a single enum Type

changed escape key from ESC to F4
removed repeated width and height in animations in gameData
created Cost struct (in Entity), changed cost methods
music is now loaded at the start of the game, via Audio_Paths.xml
removed config unused variables
removed unused methods
added more colors
add enum for building sprites (town hall is an exception)
optimized and divided commands in smaller functions
fixed hero mem leak
removed bool building in entity controller, now use to_build == NONE_ENTITY
optimized placingBuilding() method
optimized activateActionsButtons() method
comented entities time-slicing until bug review
finished paths are always removed in the same frame now
set init scene values to DEFINEs
changed Town_hall_pos var to a DEFINE
minimap now uses Color instead of SDL_Color

+++ Motor2D/UI_Text.cpp
@@ -100,7 +100,7 @@ void Text::BlitElement(bool use_camera)
 			if (hasBackground)
 			{
 				SDL_Rect rect = { globalPos.x -4 , globalPos.y - 1, section.w + 5, section.h +5};
-				Color bckg_color = { (float)background_color.r, (float)background_color.g, (float)background_color.b, (float)background_color.a };
+				Color bckg_color = { background_color.r, background_color.g, background_color.b, background_color.a };
 				App->render->DrawQuad(rect, bckg_color, true, use_camera);
 			}
 			if (outlined)
